requirements.txt
joblib==1.0.1
numpy==1.21.2
pandas==1.3.3
Pillow==8.3.2
scikit_learn==0.24.2
streamlit==0.89.0
xgboost==1.4.2
